requirements.txt
torch~=1.11.0
pillow~=9.1.0
